Move classes out of the presenter file
