+++ app/src/main/java/software/ninetofive/assignment_tests/main/SelectedScreen.kt
@@ -1,0 +1,13 @@
+package software.ninetofive.assignment_tests.main
+
+enum class SelectedScreen {
+    SCREEN_A {
+        override fun toString() = "SCREEN_A"
+    },
+    SCREEN_B {
+        override fun toString() = "SCREEN_B"
+    },
+    SCREEN_C {
+        override fun toString() = "SCREEN_C"
+    }
+}
